+++ .gitignore
@@ -24,7 +24,7 @@ conda_build_test_recipe/
 record.txt
 .vagrant/
 Vagrantfile
-env/
+/env/
 dev-init.bat
 devenv/
 Miniconda*.sh
